Chore: 벡터 덧셈 연산 분리
